creating models submodule
