+++ driver/src/withdraw_driver.rs
@@ -71,7 +71,7 @@ pub fn run_withdraw_listener<D, C>(db: &D, contract: &C) -> Result<(bool), Drive
 mod tests {
     use super::*;
     use crate::contract::tests::SnappContractMock;
-    use crate::models::tests::create_flux_for_test;
+    use crate::models::flux::tests::create_flux_for_test;
     use crate::db_interface::tests::DbInterfaceMock;
     use mock_it::Matcher::*;
     use web3::types::{H256, U256};
